api architecture changes + frames in flight
