+++ src/graphics/api/api.h
@@ -1,5 +1,5 @@
-#ifndef _GRAPHICS_SETUP_INCLUDED
-#define _GRAPHICS_SETUP_INCLUDED
+#ifndef _GRAPHICS_API_INCLUDED
+#define _GRAPHICS_API_INCLUDED
 // header includes all of setup-related headers
 // this part of code provides highly abstract interface for rendering
 // it was designed as highly configurable from outside
@@ -16,11 +16,9 @@
 #include "gpu.h"
 #include "device.h"
 #include "swapchain.h"
-#include "shaders.h"
-#include "passes.h"
-#include "pipelines.h"
+#include "shader.h"
+#include "renderpasses.h"
 #include "framebuffers.h"
-#include "cmpools.h"
-#include "cmbuffer.h"
+#include "command.h"
 
 #endif
