+++ src/graphics/render_passes/render_passes.h
@@ -1,5 +1,5 @@
-#ifndef _RENDER_PASSES_INCLUDED
-#define _RENDER_PASSES_INCLUDED
+#ifndef _GRAPHICS_RENDER_PASSESES_INCLUDED
+#define _GRAPHICS_RENDER_PASSESES_INCLUDED
 
 #include <vulkan/vulkan.h>
